feat: enable public newsletter posts (#272)
* feat: enable public newsletter rewrites

Adds a meta option to newsletter posts that allows specific newsletters to be publicly viewable on the site front-end. Currently uses the standard post template.

* feat: editor sidebar UI for toggling public meta option

* fix: better hook for disabling autosave on published newsletters

* fix: remove unnecessary args for admin_enqueue_scripts hook

* fix: fatal from using transition_post_status instead of publish hook

* feat: enable archive pages for newsletters with public posts only

Enables archive apges for Newsletter CPT, but only display posts that are marked as publicly viewable.

* feat: disable Newspack Campaigns on Newsletter CPTs

Uses the 'newspack_newsletters_assess_has_disabled_popups' filter provided by the Campaigns plugin.

* feat: set page title to "Page not found" for non-public newsletters

Previously, if a non-logged-in user tried to view a non-public newsletter post, the page would display as a 404 but the document title would still match the post title. This fixes that.

* fix: move wpseo_title filter

Moves the title filter above the template set, otherwise it won't work.

* fix: check for publish status in CC and Mailchimp

Adds a check for publish status in the Constant Contact class before sending the newsletter, to match Mailchimp (so that we don't try to send a newsletter again if it's been published already and we're just updating the post). Also restores an error message in the Mailchimp class that got lost.

* feat: disable Jetpack Related Posts module on newsletter posts

* feat: allow Posts Inserter block to be rendered on front-end

Adds a server-side render callback for the Posts Inserter so that it can show content on the front-end. Also updates the checkbox for the "disable ads" attribute to be a ToggleControl, to match the other toggleable UI elements from the plugin.

* chore: restore private status of render_mjml_component method

* chore: add @return statement to docblock

* fix: feedback from PR

- use lockPostAutosaving instead of dequeueing autosave script to disable autosave
- change language of "published as a page" to "published as a post"

* fix: send button behavior after newsletter has been sent

* feat: add settings option to enable/disable Jetpack Related Posts

Adds a checkbox option that lets editors decide whether Jetpack Related Posts should be enabled on single newsletter posts. Only shown if Related Posts module is active.