feat: simplify retrieval
With the changes in #416, separate retrieval of overlay and inline prompts is not necessary anymore, since the multiple overlays are allowed to be rendered on the page (API logic decides which is shown).
Handling of the category condition was redundant, since categorised prompts are filtered out by `should_display` method.